# api/openapi.yaml
openapi: 3.0.3
...
info:
  title: Recommendation API
  version: 1.0.0
  description: >
    API for personalized GitHub repository recommendations (behavioral
    path) and grounded semantic search over repository READMEs (RAG path).
servers:
  - url: http://localhost:8081
    description: Local development
paths:
  /v1/recommendations/{user_id}:
    get:
      summary: Get ranked recommendations for a user
      description: >
        Returns a page of ranked recommendations. Pagination uses a
        snapshot cursor: the first request (no cursor supplied) computes
        and freezes a ranked candidate pool of up to 100 items in Redis;
        subsequent requests page through that same frozen snapshot rather
        than re-ranking. Page size is fixed at 20.


        The snapshot expires after 10 minutes. A cursor that is malformed,
        expired, or references a missing session is NOT an error — it is
        treated exactly like no cursor at all, and a fresh snapshot is
        computed and returned.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^-?[0-9]+$'
          description: >
            ID of the user to fetch recommendations for. Must parse as a
            64-bit integer; anything else returns 400 invalid_user_id.
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: >
            Opaque cursor from a previous response. Omit on the first
            request to start a new ranking snapshot.
      responses:
        '200':
          description: >
            Recommendations returned successfully. A user with no active
            embedding (unknown or cold-start user) also lands here, as an
            empty page with fallback=true — not as a 404.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationsResponse'
        '400':
          description: user_id is not a valid integer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: >
            Candidate generation or snapshot storage failed (Postgres,
            Qdrant, or Redis unreachable).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/recommendations/search:
    post:
      summary: Natural-language search with grounded, cited answers
      description: >
        Retrieves relevant README chunks for a query and generates an
        answer strictly grounded in retrieved content. If no chunk clears
        the relevance threshold the LLM is never called at all — the
        response is no_results. See the `status` field.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: >
            Request processed successfully. Check `status` to distinguish
            a grounded answer from no_results or retrieval_error — all
            three are valid 200 outcomes. retrieval_error is deliberately
            a 200: the server did its job and is reporting the outcome,
            not failing to respond.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Malformed JSON body, or missing/empty query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/auth/sync:
    post:
      summary: Map an external identity onto a recsys user_id
      description: >
        Idempotent upsert keyed on external_id. Called by the frontend on
        every fresh sign-in; repeated calls for the same external_id
        return the same user_id rather than creating duplicates.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthSyncRequest'
      responses:
        '200':
          description: User synced; the recsys user_id is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSyncResponse'
        '400':
          description: Malformed JSON body, or missing/empty external_id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Upsert failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/events:
    post:
      summary: Record an engagement event
      description: >
        Publishes an engagement event to Kafka, from which the consumer
        writes it to Postgres and recomputes the user's embedding. The
        API deliberately does not write to Postgres directly: a direct
        write would persist the event but leave the user's
        recommendations frozen at their pre-event state.


        Because the write is asynchronous, a 202 means "durably queued",
        not "readable". The event will not appear in
        GET /v1/recommendations exclusions immediately.


        Idempotency: replaying a request that echoes BOTH `event_id` and
        `occurred_at` from a previous 202 is deduped and inserted once.
        Echoing only `event_id` is not sufficient — `occurred_at` is the
        events table's partition key and therefore half of the
        uniqueness key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequest'
      responses:
        '202':
          description: >
            Event accepted and acknowledged by all in-sync Kafka
            replicas. Echo both returned fields to retry safely.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: >
            Malformed JSON body, unknown event_type, non-positive
            user_id/item_id, or a non-UUID event_id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: >
            user_id or item_id does not exist. Checked up front because
            both are foreign keys: an unchecked bad id would be a
            permanent insert failure in the consumer, which drops the
            message — the client would have seen 202 for an event that
            is never stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: >
            The event could not be queued. Safe to retry with the same
            event_id and occurred_at.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: The reference check failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/items:
    get:
      summary: Browse the repository catalog
      description: >
        Lists repositories ordered by stars descending, id descending as
        a deterministic tie-break.


        Unlike GET /v1/recommendations this uses keyset pagination
        against a stored index rather than a frozen Redis snapshot: the
        ordering key is stored, not computed per user, so paging
        straight off the index is cheaper and needs no session state or
        TTL. The consequence is that a bad cursor IS an error here —
        silently restarting from the top of the catalog would look like
        working pagination while repeating items the client already
        showed.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Items per page.
        - name: language
          in: query
          required: false
          schema:
            type: string
          description: >
            Exact-match language filter, e.g. "Go". See GET /v1/languages
            for the values present in the catalog. Omit for all languages.
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: >
            Opaque cursor from a previous response. It carries the
            language filter it was created under, so `language` may be
            omitted when paging. Supplying a `language` that contradicts
            the cursor is a 400 rather than a silently mixed result set.
      responses:
        '200':
          description: A page of the catalog.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrowseResponse'
        '400':
          description: limit out of range, malformed cursor, or language contradicting the cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Listing failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/items/{item_id}:
    get:
      summary: Get one repository with its README
      description: >
        Returns repository metadata plus its README, in document order,
        as the same chunks the RAG path cites — so a client can
        cross-reference a search citation's chunk_index against this
        response.
      parameters:
        - name: item_id
          in: path
          required: true
          schema:
            type: string
            pattern: '^-?[0-9]+$'
          description: Must parse as a 64-bit integer; anything else returns 400.
      responses:
        '200':
          description: >
            The repository. A repo with no ingested README is not an
            error — `readme` is simply an empty array.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDetail'
        '400':
          description: item_id is not a valid integer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: item_id does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Fetch failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/languages:
    get:
      summary: List languages present in the catalog
      description: >
        Distinct languages ordered by repository count descending, so a
        browse UI can populate its filter control without hardcoding the
        ingest categories.
      responses:
        '200':
          description: The language list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LanguagesResponse'
        '500':
          description: Query failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    RecommendationsResponse:
      type: object
      required: [items, cursor, model_version, fallback, fallback_reason]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RecommendationItem'
        cursor:
          type: string
          nullable: true
          description: Pass back in the next request to fetch the following page. Null if no more pages.
        model_version:
          type: string
          description: >
            Version identifier of the model that produced these rankings.
            Set to "none" when fallback is true.
        fallback:
          type: boolean
          description: True if this response used degraded/non-personalized logic.
        fallback_reason:
          type: string
          nullable: true
          enum: [cold_start]
          description: >
            Populated only when fallback is true. cold_start is currently
            the only reason emitted — the user has no active embedding.
            Other degradation modes (ANN timeout, cache unavailable) are
            not yet distinguished and currently surface as 500s.
    RecommendationItem:
      type: object
      required: [item_id, score, title, description, stars]
      properties:
        item_id:
          type: string
          description: Integer item id, serialized as a string.
        score:
          type: number
          format: float
          description: >
            Blended rank score: 0.8 * ANN cosine similarity +
            0.2 * star count normalized against the candidate pool max.
        title:
          type: string
          description: >
            Repository name, without the owner prefix — the same value
            GET /v1/items returns as `name`.
        description:
          type: string
          description: Empty string when the repo has no description.
        stars:
          type: integer
          description: >
            Star count, the same figure the 0.2 weight above is
            normalized from.
    SearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
    SearchResponse:
      type: object
      required: [status, answer, citations]
      properties:
        status:
          type: string
          enum: [grounded, no_results, retrieval_error]
        answer:
          type: string
          nullable: true
          description: Null unless status is "grounded".
        citations:
          type: array
          description: >
            Chunks that cleared the relevance threshold, at most 5.
            Empty for no_results and retrieval_error.
          items:
            $ref: '#/components/schemas/Citation'
...
    AuthSyncRequest:
      type: object
      required: [external_id]
      properties:
        external_id:
          type: string
          minLength: 1
          description: >
            Stable external identity, namespaced by provider — the
            frontend sends "github:<numeric GitHub user id>".
    AuthSyncResponse:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: integer
          format: int64
    EventRequest:
      type: object
      required: [event_type, user_id, item_id]
      properties:
        event_type:
          type: string
          enum: [viewed, starred, clicked_readme]
        user_id:
          type: integer
          format: int64
          minimum: 1
        item_id:
          type: integer
          format: int64
          minimum: 1
        event_id:
          type: string
          format: uuid
          description: >
            Optional; server-generated when omitted. Supply the value
            from a previous 202 to retry that exact event.
        occurred_at:
          type: string
          format: date-time
          description: >
            Optional; defaults to the time the request is handled. When
            retrying, echo the value from the previous 202 — dedupe is
            on (event_id, occurred_at), not event_id alone.
    EventResponse:
      type: object
      required: [event_id, occurred_at, status]
      properties:
        event_id:
          type: string
          format: uuid
        occurred_at:
          type: string
          format: date-time
          description: The effective timestamp. Echo it with event_id to retry idempotently.
        status:
          type: string
          enum: [accepted]
...
    ItemDetail:
      allOf:
        - $ref: '#/components/schemas/Item'
        - type: object
          required: [readme]
          properties:
            readme:
              type: array
              description: README chunks in document order. Empty if none were ingested.
              items:
                $ref: '#/components/schemas/ReadmeSection'
    ReadmeSection:
      type: object
      required: [chunk_index, section_heading, chunk_text]
      properties:
        chunk_index:
          type: integer
          description: >
            Zero-based position within the README. Matches the
            chunk_index on a search Citation for the same repo.
        section_heading:
          type: string
          description: Empty string for content above the first heading.
        chunk_text:
          type: string
    BrowseResponse:
      type: object
      required: [items, cursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        cursor:
          type: string
          nullable: true
          description: Pass back to fetch the following page. Null on the last page.
    LanguagesResponse:
      type: object
      required: [languages]
      properties:
        languages:
          type: array
          items:
            type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string